Major refactor of plotting and estrannaise
This removes the reverse dependency, and addresses several other things:
* Changes to the model and modeldata to remove unnecessary interdependence
* Syntax cleanup
* Removal of some superfluous functions and many unneeded exports
* Removal of "core". This places all DOM interactions now into one file.
* Fixing problematic anonymous functions
