+++ src/plotting.js
@@ -1,17 +1,3 @@
-
-import {
-  colorBackground,
-  colorThePink,
-  conversionFactor,
-  convertHexToRGBA,
-  currentColorScheme,
-  menstrualCycleVisible,
-  daysAsIntervals,
-  numberToDayHour,
-  readRow,
-  getTDEs,
-  units,
-} from './core.js';
 import {
   e2MultiDose3C,
   e2ssAverage3C,
@@ -19,24 +5,56 @@ import {
   menstrualCycleP05,
   menstrualCycleP95,
   PKFunctions,
+  PKParameters,
   PKRandomFunctions,
 } from './models.js';
-import {
-  PKParams
-} from './modeldata.js';
 
 const NB_CLOUD_POINTS = 3500;
 
-const PLOT_WIDTH = 848;
-const NB_LINE_POINTS = PLOT_WIDTH;
+const NB_LINE_POINTS = 848;
 
 const CLOUD_POINT_SIZE = 1.3;
+
 const CLOUD_POINT_OPACITY = 0.4;
 
-const WONGHEXES = ['#019E73', '#E79F03', '#54ADE1', '#F0E441', '#0072B2', '#D55E00', '#CB79A7'];
+const WONG_PALETTE = [
+    'rgba(1, 158, 115, {})',   // #019E73
+    'rgba(231, 159, 3, {})',   // #E79F03
+    'rgba(84, 173, 225, {})',  // #54ADE1
+    'rgba(240, 228, 65, {})',  // #F0E441
+    'rgba(0, 114, 178, {})',   // #0072B2
+    'rgba(213, 94, 0, {})',    // #D55E00
+    'rgba(203, 121, 167, {})', // #CB79A7
+];
 
-function wongPalette(n, alpha=1.0) {
-    return convertHexToRGBA(WONGHEXES[n % WONGHEXES.length], alpha);
+function wongPalette(n, alpha = 1.0) {
+    return WONG_PALETTE[n % WONG_PALETTE.length].replace('{}', alpha);
+}
+
+function convertCustomCSSVarToRGBA(varName, alpha = 1.0) {
+    let rootStyle = getComputedStyle(document.documentElement);
+    let color = rootStyle.getPropertyValue(varName);
+    let rgb = '';
+    if (color.startsWith('#')) {
+        rgb = color;
+        let r = parseInt(rgb.slice(1, 3), 16);
+        let g = parseInt(rgb.slice(3, 5), 16);
+        let b = parseInt(rgb.slice(5, 7), 16);
+        rgb = `${r}, ${g}, ${b}`;
+    } else {
+        rgb = color.substring(4, color.length - 1);
+    }
+    return `rgba(${rgb}, ${alpha})`;
+}
+
+function colorThePink(alpha = 1.0) { return convertCustomCSSVarToRGBA('--the-pink', alpha); }
+
+function colorBackground(alpha = 1.0) { return convertCustomCSSVarToRGBA('--background-color', alpha); }
+
+function numberToDayHour(number, precision = 0) {
+    let days = Math.floor(number);
+    let hours = (Math.round((number - days) * 24)).toFixed(precision);
+    return `${days}d ${hours}h`;
 }
 
 function sum(array) {
@@ -54,42 +72,58 @@ function fillCurve(func, xmin, xmax, nbsteps) {
 function fillMenstrualCycleCurve(xmin, xmax, nbsteps) {
     let curve = [];
     for (let i = xmin; i <= xmax; i += (xmax - xmin) / (nbsteps - 1)) {
-        curve.push({ Time: i, E2: menstrualCycle(i), E2p5: menstrualCycleP05(i), E2p95: menstrualCycleP95(i)});
+        curve.push({
+            Time: i,
+            E2: menstrualCycle(options.conversionFactor, i),
+            E2p5: menstrualCycleP05(options.conversionFactor, i),
+            E2p95: menstrualCycleP95(options.conversionFactor, i)
+        });
     }
     return curve;
 
 }
 
 /**
- * Re-draw the graph and all curves
- * TODO: Take in all arguments instead of stealing them from globals or DOM
- * @return Plot for insertion into DOM
+ * Re-calculate all curves and rebuild the graph
+ * @param {Object} firstRow from multi-dose table
+ * @param {Array} multiDoses
+ * @param {Array} steadyDoses
+ * @param {Object} options containing the following keys:
+ * * `conversionFactor`: number,
+ * * `currentColorScheme`: String
+ * * `daysAsIntervals`: boolean,
+ * * `menstrualCycleVisible`: boolean,
+ * * `units`: String
+ * @return Plot for insertion into the page
  */
-export function plotCurves() {
+export function plotCurves(firstRow, multiDoses, steadyDoses, options) {
+    // track the max e2 across all multi-dose curves
+    // to set the y-axis limit. uncertainty clouds ignored.
+    let e2max = 0;
+    let xmax = 70; // FIXME: Describe what this is
+    let colorCycle = 5;
     let dotMarks  = [],
         lineMarks = [],
         ruleMarks = [],
-        tipMarks  = [];
+        tipMarks  = [],
+        msMarks   = [];
 
     // If the first row of the multi-dose table is invalid
     // we won't know about mdCVisib and mdUVisib
-    let [mdTimes, mdDoses, mdTypes, [mdCVisib, ..._cnulls], [mdUVisib, ..._unulls]] = getTDEs('multidose-table', true);
-    // So read it again with keepincomplete=true
-    let firstRow = readRow(document.getElementById('multidose-table').rows[1], true); // FIXME: Remove
-    mdCVisib = firstRow.cvisibility;
-    mdUVisib = firstRow.uvisibility;
+    let [mdTimes, mdDoses, mdTypes, [mdCVisib, ..._cnulls], [mdUVisib, ..._unulls]] = multiDoses;
+    // FIXME: Have getTDEs() more robust against this to prevent this
+    mdCVisib = firstRow.cVisibility;
+    mdUVisib = firstRow.uVisibility;
 
-    let [ssEveries, ssDoses, ssTypes, ssCVisibs, ssUVisibs] = getTDEs('steadystate-table', true);
+    let [ssEveries, ssDoses, ssTypes, ssCVisibs, ssUVisibs] = steadyDoses;
 
     let xmin = 0;
-    if (!daysAsIntervals) {
+    if (!options.daysAsIntervals) {
         xmin = Math.min(0, ...mdTimes);
     }
 
-    let xmax = 70; // FIXME: Describe what this is
-
     if (mdCVisib || mdUVisib) {
-        if (daysAsIntervals) {
+        if (options.daysAsIntervals) {
             xmax = Math.max(xmax, 1.618 * (sum(mdTimes) - (mdTimes[0] ? mdTimes[0] : 0)));
         } else {
             xmax = Math.max(xmax, 1.618 * Math.max(...mdTimes));
@@ -102,23 +136,29 @@ export function plotCurves() {
         }
     }
 
-    // track the max e2 across all multi-dose curves
-    // to set the y-axis limit. uncertainty clouds ignored.
-    let e2max = 0;
-
-    let msMarks = [];
-    if (menstrualCycleVisible) {
+    if (options.menstrualCycleVisible) {
         let _menstrualCycle = fillMenstrualCycleCurve(xmin, xmax, NB_LINE_POINTS);
         msMarks = [
-            Plot.line(_menstrualCycle, { x: 'Time', y: 'E2', strokeWidth: 2, stroke: currentColorScheme == 'night' ? convertHexToRGBA('#FFFFFF', 0.6) : convertHexToRGBA('#000000', 0.5)}),
-            Plot.areaY(_menstrualCycle, { x: 'Time', y1: 'E2p5', y2: 'E2p95' , fill: currentColorScheme == 'night' ? convertHexToRGBA('#FFFFFF', 0.1) : convertHexToRGBA('#000000', 0.1)}),
+            Plot.line(_menstrualCycle, {
+                x: 'Time',
+                y: 'E2',
+                strokeWidth: 2,
+                stroke: options.currentColorScheme == 'night' ? 'rgba(255, 255, 255, 0.6)' : 'rgba(0, 0, 0, 0.5)'}),
+            Plot.areaY(_menstrualCycle, {
+                x: 'Time',
+                y1: 'E2p5',
+                y2: 'E2p95',
+                fill: options.currentColorScheme == 'night' ? 'rgba(255, 255, 255, 0.1)' : 'rgba(0, 0, 0, 0.1)'}),
             Plot.tip(_menstrualCycle, Plot.pointerX({
-                x: 'Time', y: 'E2',
-                title: p => `menstrual cycle\ntime: ${numberToDayHour(p.Time)}\n  e₂: ${p.E2.toFixed(0)} ${units}\n  CI: ${p.E2p5.toFixed(0)}-${p.E2p95.toFixed(0)} ${units}`,
-                fontFamily: 'monospace', fill: colorBackground(0.618), stroke: colorThePink()
+                x: 'Time',
+                y: 'E2',
+                title: p => `menstrual cycle\ntime: ${numberToDayHour(p.Time)}\n  e₂: ${p.E2.toFixed(0)} ${options.units}\n  CI: ${p.E2p5.toFixed(0)}-${p.E2p95.toFixed(0)} ${options.units}`,
+                fontFamily: 'monospace',
+                fill: colorBackground(0.618),
+                stroke: colorThePink()
             }))
         ];
-        e2max = Math.max(e2max, conversionFactor * 350);
+        e2max = Math.max(e2max, options.conversionFactor * 350);
     }
 
     if (mdTimes.length > 0) {
@@ -131,37 +171,39 @@ export function plotCurves() {
             // The cloud points are not used to set it since they
             // sometimes are too far up.
             if (!mdCVisib) {
-                let probeMultiDoseCurve = fillCurve(t => e2MultiDose3C(t, mdDoses, mdTimes, mdTypes, false, daysAsIntervals), xmin, xmax, NB_LINE_POINTS);
+                let probeMultiDoseCurve = fillCurve(t => e2MultiDose3C(t, mdDoses, mdTimes, mdTypes, options.conversionFactor, false, options.daysAsIntervals), xmin, xmax, NB_LINE_POINTS);
                 e2max = Math.max(e2max, ...probeMultiDoseCurve.map(p => p.E2));
             }
 
             let mdUncertaintyCloud = [];
             for (let i = 0; i < NB_CLOUD_POINTS; i++) {
                 let randx = Math.random() * (xmax - xmin) + xmin;
-                let y = e2MultiDose3C(randx, mdDoses, mdTimes, mdTypes, true, daysAsIntervals);
+                let y = e2MultiDose3C(randx, mdDoses, mdTimes, mdTypes, options.conversionFactor, true, options.daysAsIntervals);
                 mdUncertaintyCloud.push({ Time: randx, E2: y });
             }
             dotMarks.push(Plot.dot(mdUncertaintyCloud, { x: 'Time', y: 'E2', r: CLOUD_POINT_SIZE, fill: wongPalette(4, CLOUD_POINT_OPACITY) }));
         }
 
-
         if (mdCVisib) {
-            let multiDoseCurve = fillCurve(t => e2MultiDose3C(t, mdDoses, mdTimes, mdTypes, false, daysAsIntervals), xmin, xmax, NB_LINE_POINTS);
+            let multiDoseCurve = fillCurve(t => e2MultiDose3C(t, mdDoses, mdTimes, mdTypes, options.conversionFactor, false, options.daysAsIntervals), xmin, xmax, NB_LINE_POINTS);
             multiDoseCurve = multiDoseCurve.map(p => ({ Time: p.Time, E2: p.E2 }));
 
             e2max = Math.max(e2max, ...multiDoseCurve.map(p => p.E2));
-            lineMarks.push(Plot.line(multiDoseCurve, { x: 'Time', y: 'E2', strokeWidth: 2, stroke: wongPalette(4), strokeDash: [2, 2]}));
+            lineMarks.push(Plot.line(multiDoseCurve, {
+                x: 'Time',
+                y: 'E2',
+                strokeWidth: 2, 
+                stroke: wongPalette(4), strokeDash: [2, 2]
+            }));
 
             tipMarks.push(Plot.tip(multiDoseCurve, Plot.pointerX({
                 x: 'Time', y: 'E2',
-                title: p => `multi-dose\n\ntime: ${numberToDayHour(p.Time)}\n  e₂: ${p.E2.toFixed(0)} ${units}`,
+                title: p => `multi-dose\n\ntime: ${numberToDayHour(p.Time)}\n  e₂: ${p.E2.toFixed(0)} ${options.units}`,
                 fontFamily: 'monospace', fill: colorBackground(0.618), stroke: colorThePink()
             })));
         }
 
     }
-
-    let colorCycle = 5;
 
     for (let i = 0; i < ssEveries.length; i++) {
 
@@ -170,30 +212,39 @@ export function plotCurves() {
             let ssUncertaintyCloud = [];
             for (let j = 0; j < NB_CLOUD_POINTS; j++) {
                 let randx = Math.random() * (xmax - xmin) + xmin;
-                let y = PKRandomFunctions[ssTypes[i]](randx, ssDoses[i], true, ssEveries[i]);
+                let y = PKRandomFunctions(options.conversionFactor)[ssTypes[i]](randx, ssDoses[i], true, ssEveries[i]);
                 ssUncertaintyCloud.push({ Time: randx, E2: y });
             }
 
             if (!ssCVisibs[i]) {
-                let probeSteadyStateCurve = fillCurve(t => PKFunctions[ssTypes[i]](t, ssDoses[i], true, ssEveries[i]), xmin, xmax, NB_LINE_POINTS);
+                let probeSteadyStateCurve = fillCurve(t => PKFunctions(options.conversionFactor)[ssTypes[i]](t, ssDoses[i], true, ssEveries[i]), xmin, xmax, NB_LINE_POINTS);
                 e2max = Math.max(e2max, ...probeSteadyStateCurve.map(p => p.E2));
             }
 
-            dotMarks.unshift(Plot.dot(ssUncertaintyCloud, { x: "Time", y: "E2", r: CLOUD_POINT_SIZE, fill: wongPalette(colorCycle, CLOUD_POINT_OPACITY) }));
+            dotMarks.unshift(Plot.dot(ssUncertaintyCloud, {
+                x: 'Time',
+                y: 'E2',
+                r: CLOUD_POINT_SIZE,
+                fill: wongPalette(colorCycle, CLOUD_POINT_OPACITY)
+            }));
         }
 
         if (ssCVisibs[i]) {
-            let ssEsterCurve = fillCurve(t => PKFunctions[ssTypes[i]](t, ssDoses[i], true, ssEveries[i]), xmin, xmax, NB_LINE_POINTS);
-            ssEsterCurve = ssEsterCurve.map(p => ({ Time: p.Time, E2: p.E2, type: `${ssTypes[i]} ${ssDoses[i]}mg/${ssEveries[i]}day${ssEveries[i] > 1 ? 's' : ''}` }));
+            let ssEsterCurve = fillCurve(t => PKFunctions(options.conversionFactor)[ssTypes[i]](t, ssDoses[i], true, ssEveries[i]), xmin, xmax, NB_LINE_POINTS);
+            ssEsterCurve = ssEsterCurve.map(p => ({
+                Time: p.Time,
+                E2: p.E2,
+                type: `${ssTypes[i]} ${ssDoses[i]}mg/${ssEveries[i]}day${ssEveries[i] > 1 ? 's' : ''}`
+            }));
             e2max = Math.max(e2max, ...ssEsterCurve.map(p => p.E2));
             lineMarks.unshift(Plot.line(ssEsterCurve, { x: 'Time', y: 'E2', strokeWidth: 2, stroke: wongPalette(colorCycle) }));
             tipMarks.unshift(Plot.tip(ssEsterCurve, Plot.pointerX({
                 x: 'Time', y: 'E2',
                 title: p => `${p.type.toLowerCase()},
                     time: ${numberToDayHour(p.Time)},
-                    e₂: ${p.E2.toFixed(0)} ${units},
-                    average: ${ssTypes[i].includes('patch') ? 'unavailable' : e2ssAverage3C(conversionFactor * ssDoses[i], ssEveries[i], ...PKParams[ssTypes[i]]).toFixed(0)} ${ssTypes[i].includes("patch") ? '' : units}
-                    trough: ${PKFunctions[ssTypes[i]](0.0, ssDoses[i], true, ssEveries[i]).toFixed(0)} ${units}`,
+                    e₂: ${p.E2.toFixed(0)} ${options.units},
+                    average: ${ssTypes[i].includes('patch') ? 'unavailable' : e2ssAverage3C(options.conversionFactor * ssDoses[i], ssEveries[i], ...PKParameters[ssTypes[i]]).toFixed(0)} ${ssTypes[i].includes("patch") ? '' : options.units}
+                    trough: ${PKFunctions(options.conversionFactor)[ssTypes[i]](0.0, ssDoses[i], true, ssEveries[i]).toFixed(0)} ${options.units}`,
                 fontFamily: 'monospace', fill: colorBackground(0.618), stroke: colorThePink()
             })));
         }
@@ -202,9 +253,9 @@ export function plotCurves() {
     }
 
     let e2curve = Plot.plot({
-        width: 848,
+        width: NB_LINE_POINTS,
         x: { label: 'time (days)' },
-        y: { domain: [0, 1.25 * e2max], label: `serum e₂ (${units})` },
+        y: { domain: [0, 1.25 * e2max], label: `serum e₂ (${options.units})` },
         marks: [
             Plot.gridX({ stroke: 'grey' }),
             Plot.gridY({ stroke: 'grey' }),
@@ -222,10 +273,5 @@ export function plotCurves() {
         textElement.style.fontFamily = 'monospace';
     });
 
-    // FIXME: Remove DOM manipulation, handle upstream
-    let plot = document.getElementById('plot-region');
-    plot.innerHTML = '';
-    plot.append(e2curve);
     return e2curve;
-
 }
